added blocking time to read

+++ utils/playernav/playernav.h
@@ -92,7 +92,7 @@ void fini_player(playerc_mclient_t* mclient,
 
 /* Parse command line arguments, of the form host:port */
 int parse_args(int argc, char** argv,
-               int* num_bots, char** hostnames, int* ports, double*
+               size_t* num_bots, char** hostnames, int* ports, double*
                zoom, int* aa, int* map_idx);
 
 #endif
